fix bug when agent has no candidate associations

+++ work/views.py
@@ -298,8 +298,9 @@ def share_payment(request, agent_id):
         
         #import pdb; pdb.set_trace()
         aa = agent.candidate_association()
-        if aa.has_associate == pay_to_agent:
-            aa.delete()
+        if aa:
+            if aa.has_associate == pay_to_agent:
+                aa.delete()
         
         association_type = AgentAssociationType.objects.get(name="Member")
         fc_aa = AgentAssociation(
